Creando los archivos para los stores de vuex
